DTS: Remove the core hotplug on low battery feature and Decrease BCL low battery threshold to 5%
It makes sense to make BCL kick in at a battery
level of 5% instead of 10% to avoid an unnecessary
negative performance impact.

and by default BCL is configured to hotplug cpu4
cpu5, cpu6, and cpu7 when the SOC battery
level threshold is reached.

And since we already have CPUQuiet, we don't need
to hotplug the CPU again.

This behaviour is unexpected and can result in
major performance drops.

So we reconfigure BCL to avoid hotplugging any
cores once the SOC battery level threshold
is reached.

Change-Id: I6d8f0393fc73af9b5e3e7e318b54408db9a54dfb